feat(io): update kof_io_read_range and kof_io_read_range_path to use long for length parameter; enhance file reading with offset support

+++ kof-compiler/src/main/java/dev/kof/compiler/KofIo.java
@@ -54,7 +54,7 @@ final class KofIo {
             case "readBytes" -> argCount == 0 ? new IoCall("kof_io_read_bytes", INT_ARRAY, List.of()) : null;
             // leitura com offset p/ arquivos grandes (GGUF de LLM): sem carregar o arquivo inteiro
             case "readRange" -> argCount == 2
-                    ? new IoCall("kof_io_read_range", INT_ARRAY, List.of(LONG, INT)) : null;
+                    ? new IoCall("kof_io_read_range", INT_ARRAY, List.of(LONG, LONG)) : null;
             case "writeBytes" -> argCount == 1 ? new IoCall("kof_io_write_bytes", BOOL, List.of(INT_ARRAY)) : null;
             case "appendBytes" -> argCount == 1 ? new IoCall("kof_io_append_bytes", BOOL, List.of(INT_ARRAY)) : null;
             case "delete" -> isDirectory(receiver)
